fixed bug

+++ install.sh
@@ -2,7 +2,7 @@
 
 cd ~
 sudo apt update && upgrade -y
-sudo apt install git build-essential autoconf automake gdb git libffi-dev zlib1g-dev libssl-dev tmux -y
+sudo apt install git gh build-essential autoconf automake gdb git libffi-dev zlib1g-dev libssl-dev tmux -y
 git clone https://github.com/tmux-plugins/tpm ~/.tmux/plugins/tpm
 
 echo "set -g mouse on
@@ -16,7 +16,11 @@ run '~/.tmux/plugins/tpm/tpm'" > ~/.tmux.conf
 
 sudo apt install fuse curl -y
 curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim.appimage
-chmod u+x nvim.appimage
-sudo cp nvim.appimage /usr/bin/nvim
+chmod 777 nvim.appimage
 
+if ! test -f /usr/bin/nvim; then
+    sudo cp nvim.appimage /usr/bin/nvim
+fi
+
+git clone https://github.com/NvChad/NvChad ~/.config/nvim --depth 1 && nvim
